requirements.txt
numpy==2.4.0
torchvision==0.24.1
tqdm==4.67.1
hydra-core==1.3.2
tensorboard==2.20.0